requirements.txt
functions-framework==3.*
openai
line-bot-sdk
python-dotenv